requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.0
django-allauth==0.44.0
django-cleanup==5.1.0
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.0
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
virtualenv==20.2.2
wrapt==1.12.1
